resume-screener-backend/requirements.txt

flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
pdfminer.six==20221105
python-docx==0.8.11
google-generativeai==0.3.1
scikit-learn==1.3.0
numpy==1.25.2
sentence-transformers==2.2.2
gunicorn==21.2.0
